+++ Timer.h
@@ -6,18 +6,18 @@ class Timer
     int startTime;
 
     // Keep a last time not to cache this value, but to modify it if needed.
-    int lastTime;
+    float lastTime;
 
   public:
     static const int SECOND;
 
     Timer();
 
-    void zero();             // Zero time.
-    void clamp_ms( int ms ); // Clamp  lastTime.
-    int  reset();            // Update startTime.
-    int  update();           // Update lastTime.
+    void zero();               // Zero time.
+    void clamp_ms( float ms ); // Clamp  lastTime.
+    int  reset();              // Update startTime.
+    int  update();             // Update lastTime.
 
-    int   time_ms();
+    float time_ms();
     float time_sec();
 };
